working copy

+++ internal/app/Goal.go
@@ -3,7 +3,7 @@ package main
 // Goal ..
 type Goal struct {
 	ID         uint            `db:"id"`
-	PlayerID   int64           `db:"player_id"`
+	PlayerID   uint            `db:"player_id"`
 	PlayerName string          `db:"player_name"`
 	MatchID    uint            `db:"match_id"`
 	MatchRef   CalculatedMatch `gorm:"foreignkey:match_id"`
